requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.11
django-allauth==0.41.0
django-filter==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
lml==0.0.9
oauthlib==3.1.0
openpyxl==3.0.3
pyexcel-io==0.5.20
pyexcel-xlsx==0.5.8
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
twilio==6.38.0
urllib3==1.25.8
